Move deploy_ functions into ContractDeployer
from __main__.py.

This is a part of https://github.com/raiden-network/raiden-contracts/issues/714
